fix: allow numbers to be floats instead of int only

+++ api/main.py
@@ -13,13 +13,13 @@ def post_numbers(data: List[int]):
     return {"sum": sum(data)}
 
 @app.post("/mean/")
-def post_mean(list_num: List[int]):
+def post_mean(list_num: List[float]):
     mean = sum(list_num) / len(list_num)
     mean = round(mean, 3)
     return {"mean": mean}
 
 @app.post("/stddev/")
-def post_stddev(list_num: List[int]):
+def post_stddev(list_num: List[float]):
     stddev = np.std(list_num)
     stddev = round(stddev, 3)
     return {"stddev": stddev}
